requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.9
psycopg2-binary==2.9.9
sqlparse==0.5.1
